+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the sdc11073 module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.3] - 2021-08-31
+### Changed
+- operation invocation error handling, e.g. in case the HANDLE to the operation object does not match 
+  the invocation request message
+
 ## [1.1.1] - 2021-04-22
 ### Fixed
 - fixed bug if context descriptor is updated, obsolete context states were not correctly identified and deleted.